+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 - Added devtool test `-m|--cpuset-mems` flag for memory confinement when tests
   run.
 - Added the virtio traditional memory ballooning device.
+- Added `--no-seccomp` parameter for disabling the default seccomp filters.
+- Added `--seccomp-filter` parameter for supplying user-provided, custom filters.
 
 ### Changed
 
@@ -66,6 +68,7 @@
   or the ones defined indirectly by using `--node`.
 - Changed `devtool build` to build jailer binary for `musl` only targets. Building
   jailer binary for `non-musl` targets have been removed.
+- Removed the `--seccomp-level parameter`.
 
 ## [0.22.0]